Add queue dispatcher to improve performance
Optionally, the server could be started with a predefined
number of threads in a concurrent queue which gets the
processing events and execute them in a FIFO.

By default, the threads are created on demand, but for
high loads, the queue mechanism could save time avoiding
to create and destroy the used threads.

Also, we improve build.sh script to allow prepend/export
variable values also for --auto mode.