Implement asynchronous timers to free nghttp2
In the previous implementation, those responses
with delay time didn't release the http2 io service
globbling new threads on dynamic threads creation
model (default in the server class) or freezing the
worker threads limiting the server performance.
Also, when the hardware concurrency is not enough,
dynamic threads created run out of sense growing
without control and causing overhead.

The root of the problem is that delays were performed
in a very bad way. This commit implements asynchronous
timers to free the nghttp2 io service loop. During the
delay, the worker thread is available to manage new
receptions, and even send responses directly when no
delay is requested.

Also, the dynamic threads creation model is removed,
keeping only the thread pool dispatcher for a given
fixed number of threads to reduce the threads creation
overhead, because normally a few number of threads are
enough to process everything when no delays are performed
within.

An specific io service for timers will be provided to
the server to be used by created timers. A shared
pointer is passed to the lambda async function in order
to keep alive the ownership and execute the delay during
the needed time.