Avoid queue dispatcher for 1 worker thread
It is overkilling to use the queue dispatcher if only
one server worker is requested. Indeed, this pattern
is useful when we have programmed delays for responses
and omitting it in normal situation is much more efficient.
Normally 1 worker thread is enough to manage high load.

Also, we minimize the crash with high load when breaking
client connection [0].

[0] https://github.com/testillano/h2agent/issues/6