Protect delayed response without connection
When a delay timer expires, the stream is commited.
But when there is no connection, it ended with segmentation
fault. So, we will signal the broken connection event to check
before post on response struct. New member error_ and error()
to enable it are implemented.

Stream is now a normal class, not needed to be self referenced.