Remove queue dispatcher
We remove the QueueDispatcher class from the library.
A new project has been created for it [0]

Also, we implement a simple algorithm for congestion
control, which consist in provide the 'congestion'
flag to reception() when there are no idle consumers
available to process queue items, and queue size is
greater than a maximum provided. The congestion
control is disabled by default.

[0] https://github.com/testillano/queuedispatcher