Add dynamic response delay mechanism
Add virtual server method to implement the
read of delay response (zero to cancel, which
is the default implementation) after request
reception.

So, when request is received, a programmed delay
response (receive) could be planned, and after
that, the dynamic read of next delay responses
are managed by the new virtual method:

   responseDelayTimer()

Also, remove the correction of delays, because the
differences were around few microseconds and the
calculations do not worth it.